requirements.txt
pillow
opencv_python
pyautogui
cairosvg
paddleocr
paddlepaddle
numpy
